use pkgbase to look up info for package
When you run `aura -Ai <package>`, it sends an `info` RPC to the AUR to
look up basic metadata on the package. Aura does not only display this,
though -- it fetches and parses `PKGBUILD` for more information. When it
does this, it uses the `aurNameOf` from the response object to look up
the `PKGBUILD` file, but that is the _package name_, and the AUR
respositories are organized by _package bases_, since a single package
base may contain multiple package names. We need to use the `pkgBaseOf`
instead to look up the information.